Require rust >= 1.26 and drop libc_int128 conditional
[ resolve conflicts and add test skips that seem to be needed now -
  Trevor ]

(apply <https://github.com/rust-lang/libc/pull/4060> to `main`)
(cherry picked from commit 23a0d01d9e10666096c360f2f1e8105fc0f8cfb1)

Changes are reduced in this cherry pick compared to the original commit
because some of this was already applied to `main`.
