--- src/unix/linux_like/linux/musl/b64/aarch64/int128.rs
@@ -1,7 +1,0 @@
-s! {
-    pub struct user_fpsimd_struct {
-        pub vregs: [::__uint128_t; 32],
-        pub fpsr: u32,
-        pub fpcr: u32,
-    }
-}
