+++ src/unix/linux_like/linux/musl/b64/aarch64/mod.rs
@@ -101,6 +101,12 @@ s! {
         pub set_tid_size: ::c_ulonglong,
         pub cgroup: ::c_ulonglong,
     }
+
+    pub struct user_fpsimd_struct {
+        pub vregs: [::__uint128_t; 32],
+        pub fpsr: u32,
+        pub fpcr: u32,
+    }
 }
 
 s_no_extra_traits! {
@@ -686,6 +692,3 @@ pub const VMIN: usize = 6;
 pub const IEXTEN: ::tcflag_t = 0x00008000;
 pub const TOSTOP: ::tcflag_t = 0x00000100;
 pub const FLUSHO: ::tcflag_t = 0x00001000;
-
-mod int128;
-pub use self::int128::*;
